Cleaning up Settings My Account code
