+++ src/components/Settings/SettingsInterests.js
@@ -10,7 +10,7 @@ import { useAuth } from '../../util/auth'
 import { updateUser } from '../../util/db'
 import { useTranslation } from 'react-i18next'
 
-function SettingsInterests({ setShowComponent }) {
+function SettingsInterests({ showComponent, setShowComponent }) {
   const auth = useAuth()
   const { t } = useTranslation()
 
@@ -24,62 +24,65 @@ function SettingsInterests({ setShowComponent }) {
   const clippedCategories = categories.slice(1)
 
   return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        height: '700px',
-      }}
-    >
-      <Box>
-        <Box sx={{ padding: '1.5rem 0' }}>
-          <Typography variant="h5">I'm interested in...</Typography>
-        </Box>
-        <Stack direction="row" spacing={2} useFlexGap flexWrap="wrap">
-          {clippedCategories.splice(1).map((category, index) => {
-            const clicked = interests.includes(category.label)
-            return (
-              <Chip
-                key={index}
-                label={category.label}
-                clickable
-                disabled={
-                  interests.length === 3 && !interests.includes(category.label)
-                }
-                onClick={() => handleInterests(category.label)}
-                sx={{
-                  fontSize: 16,
-                  marginLeft: 0,
-                  padding: '10px 0',
-                  backgroundColor: clicked ? '#6956F1' : '#211E34',
-                }}
-                variant="default"
-              />
-            )
-          })}
-        </Stack>
-      </Box>
+    showComponent === 'interests' && (
       <Box
         sx={{
           display: 'flex',
-          padding: '1.5rem 0',
-          flexGrow: 1,
-          alignItems: 'flex-end',
+          flexDirection: 'column',
+          height: '700px',
         }}
       >
-        <Button
-          fullWidth
-          color="info"
-          sx={{ borderRadius: '25px' }}
-          onClick={() => {
-            updateUser(auth.user.uid, { interests: interests })
-            setShowComponent('nav')
+        <Box>
+          <Box sx={{ padding: '1.5rem 0' }}>
+            <Typography variant="h5">I'm interested in...</Typography>
+          </Box>
+          <Stack direction="row" spacing={2} useFlexGap flexWrap="wrap">
+            {clippedCategories.splice(1).map((category, index) => {
+              const clicked = interests.includes(category.label)
+              return (
+                <Chip
+                  key={index}
+                  label={category.label}
+                  clickable
+                  disabled={
+                    interests.length === 3 &&
+                    !interests.includes(category.label)
+                  }
+                  onClick={() => handleInterests(category.label)}
+                  sx={{
+                    fontSize: 16,
+                    marginLeft: 0,
+                    padding: '10px 0',
+                    backgroundColor: clicked ? '#6956F1' : '#211E34',
+                  }}
+                  variant="default"
+                />
+              )
+            })}
+          </Stack>
+        </Box>
+        <Box
+          sx={{
+            display: 'flex',
+            padding: '1.5rem 0',
+            flexGrow: 1,
+            alignItems: 'flex-end',
           }}
         >
-          {t('settings.update')}
-        </Button>
+          <Button
+            fullWidth
+            color="info"
+            sx={{ borderRadius: '25px' }}
+            onClick={() => {
+              updateUser(auth.user.uid, { interests: interests })
+              setShowComponent('nav')
+            }}
+          >
+            {t('settings.update')}
+          </Button>
+        </Box>
       </Box>
-    </Box>
+    )
   )
 }
