+++ src/components/Settings/SettingsMyAccount.js
@@ -1,15 +1,8 @@
 import React, { useState } from 'react'
 import Container from '@mui/material/Container'
-import List from '@mui/material/List'
-import ListItem from '@mui/material/ListItem'
-import Typography from '@mui/material/Typography'
-import Grid from '@mui/material/Grid'
-import ListItemText from '@mui/material/ListItemText'
-import ListItemButton from '@mui/material/ListItemButton'
-import ChevronRightIcon from '@mui/icons-material/ChevronRight'
-import EastIcon from '@mui/icons-material/East'
 
 import ArrowBack from '../ArrowBack'
+import MyAccountNav from './SettingsMyAccountNav'
 import SettingsInterests from './SettingsInterests'
 import SettingsDisplayName from './SettingsDisplayName'
 import SettingsCommunity from './SettingsCommunity'
@@ -18,7 +11,6 @@ import SettingsLanguage from './SettingsLanguage'
 import SettingsSchools from './SettingsSchools'
 import SettingsDeleteAccount from './SettingsDeleteAccount'
 
-import { useTranslation } from 'react-i18next'
 import { useAuth } from '../../util/auth'
 import { requireAuth } from '../../util/auth'
 
@@ -27,6 +19,8 @@ function SettingsMyAccount(props) {
 
   const [showComponent, setShowComponent] = useState('nav')
 
+  //ToDo: Make showComponent a custom hook so that we don't have to pass it in like this to each component.
+
   return (
     <>
       <ArrowBack
@@ -34,115 +28,45 @@ function SettingsMyAccount(props) {
         setShowComponent={setShowComponent}
       />
       <Container>
-        {showComponent === 'displayName' && (
-          <SettingsDisplayName
-            auth={auth}
-            setShowComponent={setShowComponent}
-          />
-        )}
-
-        {showComponent === 'email' && (
-          <SettingsEmail auth={auth} setShowComponent={setShowComponent} />
-        )}
-
-        {showComponent === 'school' && (
-          <SettingsSchools setShowComponent={setShowComponent} />
-        )}
-
-        {showComponent === 'interests' && (
-          <SettingsInterests setShowComponent={setShowComponent} />
-        )}
-        {showComponent === 'language' && (
-          <SettingsLanguage setShowComponent={setShowComponent} />
-        )}
-
-        {showComponent === 'communities' && (
-          <SettingsCommunity setShowComponent={setShowComponent} />
-        )}
-        {showComponent === 'nav' && (
-          <MyAccountNav auth={auth} setShowComponent={setShowComponent} />
-        )}
-        {showComponent === 'deleteAccount' && (
-          <SettingsDeleteAccount
-            auth={auth}
-            setShowComponent={setShowComponent}
-          />
-        )}
+        <MyAccountNav
+          auth={auth}
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsDisplayName
+          auth={auth}
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsEmail
+          auth={auth}
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsSchools
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsInterests
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsLanguage
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsCommunity
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
+        <SettingsDeleteAccount
+          auth={auth}
+          showComponent={showComponent}
+          setShowComponent={setShowComponent}
+        />
       </Container>
     </>
   )
 }
 
 export default requireAuth(SettingsMyAccount)
-
-function MyAccountNav({ setShowComponent, auth }) {
-  const { t } = useTranslation()
-
-  const myAccountLinks = [
-    {
-      id: 'displayName',
-      title: 'Display Name',
-      userInfo: auth?.user?.displayName ?? auth.user?.name,
-    },
-    { id: 'email', title: 'Email', userInfo: auth?.user?.email },
-    { id: 'school', title: 'School', userInfo: auth?.user?.school },
-    {
-      id: 'interests',
-      title: 'Interests',
-      userInfo: auth?.user?.interests.join(','),
-    },
-    {
-      id: 'language',
-      title: 'Language',
-      userInfo: auth?.user?.language && auth?.user?.language.join(','),
-    },
-    {
-      id: 'communities',
-      title: 'Communities',
-      userInfo: auth?.user?.fnmi && auth?.user?.fnmi.join(','),
-    },
-  ]
-
-  return (
-    <>
-      <List
-        sx={{ width: '100%', maxWidth: 450 }}
-        component="nav"
-        aria-labelledby="settings-my-account"
-      >
-        {myAccountLinks.map((accLink) => (
-          <ListItem
-            onClick={() => {
-              setShowComponent(accLink.id)
-            }}
-            key={accLink.title}
-            secondaryAction={<ChevronRightIcon />}
-          >
-            <ListItemButton>
-              <Grid container>
-                <Grid item xs={accLink.userInfo ? 7 : 12}>
-                  <Typography>{accLink.title}</Typography>
-                </Grid>
-                {accLink.userInfo && (
-                  <Grid item xs={5}>
-                    <Typography noWrap> {accLink.userInfo}</Typography>
-                  </Grid>
-                )}
-              </Grid>
-            </ListItemButton>
-          </ListItem>
-        ))}
-        <ListItem>
-          <ListItemButton
-            onClick={() => {
-              setShowComponent('deleteAccount')
-            }}
-          >
-            <ListItemText>{t('settings.delete-account')}</ListItemText>
-            <EastIcon />
-          </ListItemButton>
-        </ListItem>
-      </List>
-    </>
-  )
-}
